requirements.txt
tencentcloud-sdk-python-cvm
tencentcloud-sdk-python-vpc